+++ service/src/test/kotlin/fi/espoo/evaka/document/archival/ArchiveChildDocumentServiceTest.kt
@@ -5,8 +5,6 @@
 package fi.espoo.evaka.document.archival
 
 import fi.espoo.evaka.document.*
-import fi.espoo.evaka.document.archival.DocumentMetadataUtils.createDocumentMetadata
-import fi.espoo.evaka.document.archival.DocumentMetadataUtils.marshalMetadata
 import fi.espoo.evaka.document.childdocument.*
 import fi.espoo.evaka.identity.ExternalIdentifier
 import fi.espoo.evaka.placement.PlacementType
